# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample documentation for the SHA project (FrodidApp)",
    "version": "1.0.0",
    "title": "FrodidApp",
    "termsOfService": "",
    "contact": {
      "email": "semerkhet40@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ec2-18-185-66-15.eu-central-1.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/countries": {
      "post": {
        "tags": [
          "countries"
        ],
        "summary": "Creates new country",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Countries"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": [
          "countries"
        ],
        "summary": "Gets a list of all countries",
        "description": "This endpoint uses a get request to retrieve all countries",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "items_per_page",
            "type": "integer",
            "required": false
          },
          {
            "in": "query",
            "name": "page_number",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/countries/{id}": {
      "get": {
        "tags": [
          "countries"
        ],
        "summary": "Gets a single country",
        "description": "This endpoint uses a get request to retrieve a specific country",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country unique identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Countries"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "countries"
        ],
        "summary": "Updates a country in the store with form data",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of country that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              },
              "example": {
                "name": "Banana Republic"
              }
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "countries"
        ],
        "summary": "Deletes a country",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Country id to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Country not found"
          }
        }
      }
    },
...
    "/addresses": {
      "post": {
        "tags": [
          "addresses"
        ],
        "summary": "Creates new address",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created address object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Addresses"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": [
          "addresses"
        ],
        "summary": "Gets a list of all addresses",
        "description": "This endpoint uses a get request to retrieve all addresses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "items_per_page",
            "type": "integer",
            "required": false
          },
          {
            "in": "query",
            "name": "page_number",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/addresses/{id}": {
      "get": {
        "tags": [
          "addresses"
        ],
        "summary": "Gets a single address",
        "description": "This endpoint uses a get request to retrieve a specific address",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Address unique identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Addresses"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "addresses"
        ],
        "summary": "Updates a user in the addresses table",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of address that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created address object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Addresses"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "addresses"
        ],
        "summary": "Deletes an address",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Address id to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Address not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Countries": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Greece"
        }
      }
    },
...
    "Addresses": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string",
          "example": "Stournari"
        },
        "street_no": {
          "type": "string",
          "example": "15A"
        },
        "region": {
          "type": "string",
          "example": "Exarchia"
        },
        "zipcode": {
          "type": "string",
          "example": "75 254A"
        },
        "country_id": {
          "type": "integer",
          "example": 1
        },
        "state_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
  },
...
}